// src/types.ts
export interface QuizData {
  kgToLose?: string;
  bodyType?: string;
  targetZone?: string;
  name?: string;
  weightImpact?: string;
  happyWithAppearance?: string;
  obstacle?: string;
  benefits?: string[];
  knowsMethod?: boolean;
  currentWeight?: number;
  height?: number;
  goalWeight?: number;
  waterIntake?: string;
}